courses can now be searched
searchbar works, searchResults works. I added something to limit the
number of courses that are displayed. Main issue now is that we need to
search for unique ClassID and Teacher Combination. I'm thinking we make
a separate JSON file that only includes unique ClassID and another that
only includes unique ClassID/Teacher combinations, so the user can have
the choice to sort by teacher and classid or just classid. (useful if
user doesn't know what teacher is good and just wants the best schedule)